PAAC-478: Merged Group1P1Calculator into Group2P1Calculator, and renamed to Period1Calculator. Updated period factory class as appropriate.
