+++ test/calculators/periods/Period1CalculatorSpec.scala
@@ -23,7 +23,7 @@ import org.scalatest.prop._
 import org.scalacheck.Gen
 import calculators.results.BasicCalculator
 
-class Group2P1CalculatorSpec extends UnitSpec {
+class Period1CalculatorSpec extends UnitSpec {
 
   trait TestFixture {
     val annualAllowance = 80000L
@@ -37,7 +37,7 @@ class Group2P1CalculatorSpec extends UnitSpec {
       implicit val contribution = Contribution(2015, 123L)
 
       // test
-      val result = Group2P1Calculator().year2014CCF
+      val result = Period1Calculator().year2014CCF
 
       // check
       result shouldBe 0L
@@ -49,7 +49,7 @@ class Group2P1CalculatorSpec extends UnitSpec {
       implicit val contribution = Contribution(2015, 123L)
 
       // test
-      val result = Group2P1Calculator().year2014CCF
+      val result = Period1Calculator().year2014CCF
 
       // check
       result shouldBe 123L
@@ -63,7 +63,7 @@ class Group2P1CalculatorSpec extends UnitSpec {
       implicit val contribution = Contribution(2015, 0L).copy(amounts=Some(InputAmounts(triggered=Some(true))))
 
       // test
-      val result = Group2P1Calculator().dbist
+      val result = Period1Calculator().dbist
 
       // check
       result shouldBe 0L
@@ -77,7 +77,7 @@ class Group2P1CalculatorSpec extends UnitSpec {
       implicit val contribution = Contribution(2015, 0L).copy(amounts=Some(InputAmounts(triggered=Some(true),moneyPurchase=Some(2100000))))
 
       // test
-      val result = Group2P1Calculator().exceedingAllowance
+      val result = Period1Calculator().exceedingAllowance
 
       // check
       result shouldBe 0L
@@ -91,7 +91,7 @@ class Group2P1CalculatorSpec extends UnitSpec {
       implicit val contribution = Contribution(2015, 0L)
 
       // test
-      val result = Group2P1Calculator().preFlexiSavings
+      val result = Period1Calculator().preFlexiSavings
 
       // check
       result shouldBe 0L
@@ -105,7 +105,7 @@ class Group2P1CalculatorSpec extends UnitSpec {
       implicit val contribution = Contribution(2015, 0L).copy(amounts=Some(InputAmounts(triggered=Some(true),moneyPurchase=Some(2100000))))
 
       // test
-      val result = Group2P1Calculator().unusedAAA
+      val result = Period1Calculator().unusedAAA
 
       // check
       result shouldBe 3000000L
@@ -119,7 +119,7 @@ class Group2P1CalculatorSpec extends UnitSpec {
       implicit val contribution = Contribution(2015, 0L).copy(amounts=Some(InputAmounts(triggered=Some(true),definedBenefit=Some(1),moneyPurchase=Some(1))))
 
       // test
-      val result = Group2P1Calculator().preFlexiSavings
+      val result = Period1Calculator().preFlexiSavings
 
       // check
       result shouldBe 0L
@@ -133,7 +133,7 @@ class Group2P1CalculatorSpec extends UnitSpec {
       implicit val contribution = Contribution(2015, 0L).copy(amounts=Some(InputAmounts(triggered=Some(true),moneyPurchase=Some(2100000))))
 
       // test
-      val result = Group2P1Calculator().aaCCF
+      val result = Period1Calculator().aaCCF
 
       // check
       result shouldBe 3000000L
