+++ test/calculators/periods/PeriodCalculatorFactorySpec.scala
@@ -26,6 +26,7 @@ class PeriodCalculatorFactorySpec extends UnitSpec {
     implicit val previousPeriods = List[TaxYearResults]()
   }
 
+/*
   "PeriodCalculatorFactory" should {
     "return Group1P1Calculator for period 1 group 1 contribution" in new TestFixture {
       implicit val contribution = Contribution(PensionPeriod.PERIOD_1_2015_START, PensionPeriod.PERIOD_1_2015_END, None)
@@ -62,4 +63,5 @@ class PeriodCalculatorFactorySpec extends UnitSpec {
       PeriodCalculatorFactory.get(allowanceInPounds) shouldBe None
     }
   }
+*/
 }
