+++ test/calculators/results/Year2015Period1CalculatorSpec.scala
@@ -106,7 +106,7 @@ class Year2015Period1CalculatorSpec extends UnitSpec with GeneratorDrivenPropert
         val results = Year2015Period1Calculator.summary(Seq[TaxYearResults](), contribution)
 
         // check it
-        withClue(s"Contributions with date '$taxDay/${taxMonth+1}/$taxYear' should be supported but") { results shouldBe Some(ExtendedSummaryFields(0L,0L,8000000L,4000000L,8000000L,4000000L,0)) }
+        withClue(s"Contributions with date '$taxDay/${taxMonth+1}/$taxYear' should be supported but") { results shouldBe Some(ExtendedSummaryFields(0,0,8000000,4000000,8000000,4000000,0,0,0,0,0,0,0,0,0,5000,0,0,5000,0,false,5995000,0)) }
       }
     }
 
@@ -117,7 +117,7 @@ class Year2015Period1CalculatorSpec extends UnitSpec with GeneratorDrivenPropert
                                                                               PensionPeriod(2015, 4, 9),
                                                                               Some(InputAmounts(0L))))
         // check it
-        results shouldBe Some(ExtendedSummaryFields(0,0,8000000,4000000,8000000,4000000,0))
+        results shouldBe Some(ExtendedSummaryFields(0,0,8000000,4000000,8000000,4000000,acaCF=6000000))
       }
 
       "return annual allowance of 8000000 pence for all valid amounts" in {
